requirements.txt
numpy
pycryptodome
py_ecc
